Fetch registry and storage data from the one hook
