+++ src/pages/App.tsx
@@ -7,9 +7,8 @@ import styled from 'styled-components'
 import { useWeb3React } from '@web3-react/core'
 import useWordpressInfo from 'hooks/useWordpressInfo'
 import useDomainInfo from 'hooks/useDomainInfo'
-import useStorageInfo from 'hooks/useStorageInfo'
 import { useAppState } from 'state/application/hooks'
-import { retrieveDomainData, updateAppData } from 'state/application/actions'
+import { retrieveDomainData } from 'state/application/actions'
 import Loader from 'components/Loader'
 import Panel from './Panel'
 import Connection from './Connection'
@@ -105,16 +104,12 @@ export default function App() {
     }
   }, [chainId, domainDataTrigger, wordpressData])
 
-  const { data: domainData, isLoading: domainLoading } = useDomainInfo(domainDataTrigger)
-  const { data: storageData, isLoading: storageLoading } = useStorageInfo()
+  const { data: domainData, isLoading } = useDomainInfo(domainDataTrigger)
 
   useEffect(() => {
+    //@ts-ignore
     dispatch(retrieveDomainData(domainData ? { ...domainData } : domainData))
-  }, [domainData, domainLoading, dispatch])
-
-  useEffect(() => {
-    dispatch(updateAppData(storageData ? { ...storageData } : storageData))
-  }, [storageData, storageLoading, dispatch])
+  }, [domainData, isLoading, dispatch])
 
   const { admin, factory, router, projectName } = useAppState()
 
@@ -127,12 +122,6 @@ export default function App() {
   const appManagement = useSelector<AppState, AppState['application']['appManagement']>(
     (state) => state.application.appManagement
   )
-
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    setLoading(domainLoading || storageLoading)
-  }, [domainLoading, storageLoading])
 
   return (
     <Suspense fallback={null}>
@@ -147,7 +136,7 @@ export default function App() {
         <Web3ReactManager>
           <Popups />
 
-          {loading ? (
+          {isLoading ? (
             <LoaderWrapper>
               <Loader size="2.8rem" />
             </LoaderWrapper>
