+++ src/state/application/actions.ts
@@ -1,6 +1,6 @@
 import { createAction } from '@reduxjs/toolkit'
 import { TokenList } from '@uniswap/token-lists'
-import { StorageState } from './reducer'
+import { DomainData } from './reducer'
 
 export type PopupContent =
   | {
@@ -39,22 +39,7 @@ export enum ApplicationModal {
 
 export const setAppManagement = createAction<{ status: boolean }>('application/setAppManagement')
 
-export const retrieveDomainData = createAction<null | {
-  admin: string
-  factory: string
-  router: string
-  storageAddr: string
-  pairHash: string
-  feeRecipient: string
-  protocolFee?: number
-  totalFee?: number
-  allFeeToProtocol?: boolean
-  possibleProtocolPercent?: string[]
-  totalSwaps: string
-  disableSourceCopyright?: boolean
-}>('application/retrieveDomainData')
-
-export const updateAppData = createAction<null | StorageState>('application/updateAppData')
+export const retrieveDomainData = createAction<null | DomainData>('application/retrieveDomainData')
 
 export const updateBlockNumber = createAction<{ chainId: number; blockNumber: number }>('application/updateBlockNumber')
